requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-ckeditor==6.5.1
django-extensions==3.2.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
gunicorn==20.1.0
idna==3.4
Pillow==9.4.0
pydotplus==2.0.2
pyparsing==3.0.9
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
stripe==5.2.0
urllib3==1.26.14
